reorganize TFKeras classes
- moved some utility functions to perform/roms/tfKerasFuncs.py
- new class autoencoderTFKeras
	- parent class: autoencoderProjROM
	- child classes: autoencoderGalerkinProjTFKeras, etc.
